fixed: /root/kbengine//kbe/src/lib/network/encryption_filter.hpp:85: error: no matching function for call to 'KBEngine::Mercury::BlowfishFilter::BlowfishFilter(const std::basic_string<char, std::char_traits<char>, std::allocator<char> >&)' /root/kbengine//kbe/src/lib/network/encryption_filter.hpp:69: note: candidates are: KBEngine::Mercury::BlowfishFilter::BlowfishFilter() /root/kbengine//kbe/src/lib/network/encryption_filter.hpp:69: note:                 KBEngine::Mercury::BlowfishFilter::BlowfishFilter(const KBEngine::Mercury::BlowfishFilter&) make[2]: *** [Hybrid64/loginapp.o] Error 1

+++ kbe/src/lib/cstdkbe/blowfish.cpp
@@ -108,7 +108,7 @@ int KBEBlowfish::encrypt( const unsigned char * src, unsigned char * dest,
 	{
 		CRITICAL_MSG(boost::format("Blowfish::encrypt: "
 			"Input length (%1%) is not a multiple of block size (%2%)\n") %
-			length % (BLOCK_SIZE));
+			length % (int)(BLOCK_SIZE));
 	}
 
 	uint64 * pPrevBlock = NULL;
@@ -138,7 +138,7 @@ int KBEBlowfish::decrypt( const unsigned char * src, unsigned char * dest,
 	{
 		WARNING_MSG(boost::format("Blowfish::decrypt:"
 			"Input stream size (%1%) is not a multiple of the block size (%2%)\n") %
-			length % (BLOCK_SIZE));
+			length % (int)(BLOCK_SIZE));
 
 		return -1;
 	}
